requirements.txt
grpc-interceptor==0.15.0
grpcio-tools==1.51.1
locust==2.14.2
